requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
Django==4.1.7
django-bootstrap-datepicker-plus==5.0.4
django-filter==23.2
Pillow==10.0.0
python-dateutil==2.8.2
pydantic==2.1.1
pydantic_core==2.4.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
